atributo cpf adicionado ao cliente

+++ src/main/java/com/sicredi/challenge/service/client/DeleteClientService.java
@@ -1,5 +1,6 @@
 package com.sicredi.challenge.service.client;
 
+import com.sicredi.challenge.infra.exception.ExceptionDto;
 import com.sicredi.challenge.model.ClientModel;
 import com.sicredi.challenge.repository.ClientRepository;
 import org.springframework.http.ResponseEntity;
@@ -18,6 +19,10 @@ public class DeleteClientService {
     @Transactional
     public ResponseEntity execute(Long id) {
         ClientModel model = clientRepository.getReferenceById(id);
+        if(!model.getAgendas().isEmpty()) {
+            return ResponseEntity.badRequest()
+                    .body(new ExceptionDto(null, "O cliente está atrelado a uma pauta votada. Não pode ser excluído!"));
+        }
         clientRepository.delete(model);
         return ResponseEntity.noContent().build();
     }
